Align Temperature Signals
As of today we have a mix of float and int16 for temperature signals,
with no real rationale for the difference. This PR changes the int16 signals
to instead use float, so after this change all signals with Celisus unit use float.

Only intended for VSS 5.0 as this is a backward incompatible change.

Signed-off-by: Erik Jaegervall <erik.jaegervall@se.bosch.com>